# docs/backend/swagger.yml
swagger: "2.0"
info:
  title: 50thBeers
  description: 50th Beers blog API
  contact:
    email: root.spyro@gmail.com
  license:
    name: MIT
    url: https://choosealicense.com/licenses/mit/
  version: 1.0.0
host: api.example.com
basePath: /v1
securityDefinitions:
  JWT:
    type: apiKey
    name: Authorization
    in: header
paths: 
...
  /locations:
    get:
      summary: list of locations
      tags:
        - Locations
      parameters:
        - name: locationName
          in: query
          type: string
          required: false
          
        - name: status
          in: query
          type: string
          required: false
          
        - name: limit
          in: query
          type: number
          required: false
        
        - name: offset
          in: query
          type: number
          required: false
          
        - name: orderBy
          in: query
          type: string
          required: false
          
      responses:
        200:
          description: ok
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: true
              data:
                type: object
                properties:
                  itemsFound:
                    type: number
                    example: 1
                  items:
                    type: array
                    items:
                      $ref: "#/definitions/locationObject"
        500:
          description: server error
          schema:
            $ref: "#/definitions/serverError"
    
    post:
      summary: create a new location
      security:
        - JWT: []
      tags:
        - Locations
      parameters:
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              locationName:
                type: string
                example: Homo Sibaris
              mapsLink:
                type: string
                example: https://www.google.com/maps/embed?pb=!1m14!1m8!1m3!1d11975.595085269204!2d2.1388206!3d41.376284!3m2!1i1024!2i768!4f13.1!3m3!1m2!1s0x12a4988421daba0d%3A0xf1efff1748822f9c!2sHomo%20Sibaris!5e0!3m2!1ses!2ses!4v1699977760682!5m2!1ses!2ses
              comments:
                type: string
                example: A good place to drink artesanal beer
      responses:
        201:
          description: created
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: true
              data:
                $ref: "#/definitions/locationObject"
        409:
          description: item already exist
          schema:
            $ref: "#/definitions/conflict"
            
        500:
          description: server error
          schema:
            $ref: "#/definitions/serverError"
  /locations/{locationId}:
    get:
      summary: get single location
      parameters:
        - name: locationId
          in: path
          type: string
          required: true
      tags:
        - Locations
      responses:
        200: 
          description: ok
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: true
              data:
                $ref: "#/definitions/locationObject"
                
        404:
          description: not found
          schema:
            $ref: "#/definitions/notFound"
            
        500:
          description: server error
          schema:
            $ref: "#/definitions/serverError"
            
    patch:
      summary: update a location
      security:
        - JWT: []
      parameters:
        - name: locationId
          type: string
          in: path
          required: true
          
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              locationName:
                type: string
                example: Homo Sibaris
              mapsLink:
                type: string
                example: https://www.google.com/maps/embed?pb=!1m14!1m8!1m3!1d11975.595085269204!2d2.1388206!3d41.376284!3m2!1i1024!2i768!4f13.1!3m3!1m2!1s0x12a4988421daba0d%3A0xf1efff1748822f9c!2sHomo%20Sibaris!5e0!3m2!1ses!2ses!4v1699977760682!5m2!1ses!2ses
              comments:
                type: string
                example: A good place to drink artesanal beer
      tags:
        - Locations
      responses:
        200:
          description: ok
          schema:
            type: object
            properties:
              success: 
                type: boolean
                example: true
              data:
                $ref: "#/definitions/locationObject"
                
        404:
          description: not found
          schema: 
            $ref: "#/definitions/notFound"
            
        409:
          description: item alrady exist
          schema:
            $ref: "#/definitions/conflict"
            
        500:
          description: server error
          schema:
            $ref: "#/definitions/serverError"
            
    delete:
      summary: delete a location
      security:
        - JWT: []
      parameters:
        - name: locationId
          in: path
          type: string
          required: true
      tags:
        - Locations
      responses:
        200:
          description: ok
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: true
              data:
                type: string
                example: item successfully deleted
                
        404:
          description: not found
          schema:
            $ref: "#/definitions/notFound"
            
        500:
          description: server error
          schema:
            $ref: "#/definitions/serverError"
          
          
  
  # -----------------------------------
  #              POSTS
  # -----------------------------------
...
definitions:
...
  notFound:
    description: 404 not found
    type: object
    properties:
      success:
        type: boolean
        example: false
      data:
        type: string
        example: item not found
  conflict:
    description: 409 conflict
    type: object
    properties:
      success:
        type: boolean
        example: false
      data:
        type: string
        example: Item already exist
  serverError:
    description: 500 Internal Server Error
    type: object
    properties:
      success:
        type: boolean
        example: false
      data:
        type: string
        example: Something went wrong
...
  locationObject:
    type: object
    properties:
      id:
        type: string
        example: 1a79a4d60de6718e8e5b326e338ae533
      locationName:
        type: string
        example: 2D2Dspuma
      mapsLink:
        type: string
        example: "https://www.google.com/maps/embed?pb=!1m14!1m8!1m3!1d11967.085193103783!2d2.179417!3d41.422483!3m2!1i1024!2i768!4f13.1!3m3!1m2!1s0x12a4bd2b62a010fb%3A0x8ed384a43e80fd1b!2sCervecer%C3%ADa%202D2Dspuma%20-%20Cervecer%C3%ADa!5e0!3m2!1ses!2ses!4v1699976932448!5m2!1ses!2ses"
      createdAt:
        type: string
        example: 2023 - 10 - 10
      publicatedAt:
        type: string
        example: 2023 - 10 - 10
      updatedAT:
        type: string
        example:
      comments:
        type: string
        example: My favorite pub
      status:
        type: string
        example: PUBLIC
...